deletar usuario

+++ database/CRUD/index.php
@@ -30,7 +30,7 @@ if($sql -> rowCount() > 0) {
             <td><?= $usuario['email'];?></td>
             <td>
                 <a href="editar.php?id=<?=$usuario['id'];?>">[ EDITAR ]</a>
-                <a href="excluir.php?id=<?=$usuario['id'];?>">[ EXCLUIR ]</a>
+                <a href="excluir.php?id=<?=$usuario['id'];?>" onclick="return confirm('Tem certeza que deseja excluir?')">[ EXCLUIR ]</a>
             </td>
         </tr>
     <?php endforeach; ?>
